Main menu logic is done, with a new choice look. PixelRain is now moved into an own class.

+++ src/tetrix/sound/GameMusic.java
@@ -34,7 +34,7 @@ public class GameMusic {
 		gameMusic.getVolume();
 	}
 
-	public void setGameMusicVolume(int value){
+	public void setGameMusicVolume(float value){
 		gameMusic.setVolume(value);
 	}
 }
